+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Sort result of `kafkactl get topics`
+- `consume` now uses a simpler consumer without consumerGroup.
 
 ### Fixed
 - Do not auto-create topics when using `describe`, `consume` or `produce`